функция formatDateToLocalTimeZone  вынесена в utils
